move filtered table to sis
